Got it generally running.
Uinput is tricky, when it creates a device it assumes that the device is
in portrait mode, and it remaps the x and y axis when the orientation is
changed.  This is undesirable as Miracast and MHL tend to lock rotation,
so I simply map the coordinates myself depending on the orientation,
assuming the touchscreen always expects coordinates in  Rotation_0.

Still some bugs to fix regarding opening a null stream at times.   Also
need to decouple the calibration activity from the service and give it
its own app, which will probably solve some issues.